requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
aspose.slides==25.7.0; python_version < '3.14' and python_version >= '3.5'
attrs==25.3.0; python_version >= '3.8'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
cloudinary==1.44.1
colorama==0.4.6; platform_system == 'Windows'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==5.2.5; python_version >= '3.10'
django-cors-headers==4.7.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
ffmpeg-python==0.2.0
flatbuffers==25.2.10
future==1.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==3.2.4; python_version >= '3.9'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.10; python_version >= '3.6'
imageio==2.37.0; python_version >= '3.9'
imageio-ffmpeg==0.6.0; python_version >= '3.9'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
lazy-loader==0.4; python_version >= '3.7'
llvmlite==0.44.0; python_version >= '3.10'
lxml==6.0.1; python_version >= '3.8'
mpmath==1.3.0
mssql-django==1.6
networkx==3.5; python_version >= '3.11'
numba==0.61.2; python_version >= '3.10'
numpy==2.2.6; python_version >= '3.10'
onnxruntime==1.22.1; python_version >= '3.10'
opencv-python-headless==4.12.0.88; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
pdf2image==1.17.0
pillow==11.3.0; python_version >= '3.9'
platformdirs==4.3.8; python_version >= '3.9'
playwright==1.54.0; python_version >= '3.9'
pooch==1.8.2; python_version >= '3.7'
protobuf==6.32.0; python_version >= '3.9'
pyee==13.0.0; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pymatting==1.1.14; python_version >= '3'
pyodbc==5.2.0; python_version >= '3.8'
pyotp==2.9.0; python_version >= '3.7'
pyreadline3==3.5.4; python_version >= '3.8' and sys_platform == 'win32'
python-decouple==3.8
python-pptx==1.0.2; python_version >= '3.8'
pytz==2025.2
referencing==0.36.2; python_version >= '3.9'
rembg==2.0.67; python_version >= '3.10' and python_version < '3.14'
requests==2.32.5; python_version >= '3.9'
rpds-py==0.27.0; python_version >= '3.9'
scikit-image==0.25.2; python_version >= '3.10'
scipy==1.16.1; python_version >= '3.11'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.3; python_version >= '3.8'
sympy==1.14.0; python_version >= '3.9'
tifffile==2025.6.11; python_version >= '3.11'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.2; sys_platform == 'win32'
urllib3==2.5.0; python_version >= '3.9'
whitenoise==6.9.0; python_version >= '3.9'
xlsxwriter==3.2.5; python_version >= '3.8'
